\name{svWidgets-package}
\alias{svWidgets-package}
\alias{svWidgets}
\docType{package}

\title{ SciViews GUI API - Widgets & Windows }
\description{
  High level management of widgets, windows and other graphical resources.
}
\details{
  \tabular{ll}{
    Package: \tab svWidgets\cr
    Type: \tab Package\cr
    Version: \tab 0.9-44\cr
    Date: \tab 2013-11-30\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API widgets and windows }
